Some refactoring for tree logic re-use.
