+++ recaf-ui/src/main/java/me/coley/recaf/ui/control/tree/item/WorkspaceRootItem.java
@@ -3,6 +3,7 @@ package me.coley.recaf.ui.control.tree.item;
 import me.coley.recaf.code.ClassInfo;
 import me.coley.recaf.code.DexClassInfo;
 import me.coley.recaf.code.FileInfo;
+import me.coley.recaf.ui.control.tree.WorkspaceTreeWrapper;
 import me.coley.recaf.ui.util.Lang;
 import me.coley.recaf.workspace.Workspace;
 import me.coley.recaf.workspace.WorkspaceListener;
@@ -11,7 +12,7 @@ import me.coley.recaf.workspace.resource.*;
 import java.util.*;
 
 /**
- * Root item for {@link me.coley.recaf.ui.control.tree.WorkspaceTree}.
+ * Root item for {@link WorkspaceTreeWrapper}.
  *
  * @author Matt Coley
  */
